добавила тест с порядком выполнения
