--- src/test/java/org/example/HomePage.java
@@ -1,20 +1,0 @@
-package org.example;
-
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
-import org.openqa.selenium.support.FindBy;
-import org.openqa.selenium.support.PageFactory;
-
-public class HomePage {
-    public WebDriver driver;
-    public HomePage(WebDriver driver) {
-        PageFactory.initElements(driver, this);
-        this.driver = driver; }
-
-    @FindBy(className="title")
-    private WebElement pageTitle;
-
-    public String getPageTitle(){
-        return pageTitle.getText();
-    }
-}
