+++ src/test/java/org/example/helpers/LoginPage.java
@@ -1,0 +1,32 @@
+package org.example.helpers;
+
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.support.FindBy;
+import org.openqa.selenium.support.PageFactory;
+
+public class LoginPage {
+
+    public WebDriver driver;
+    public LoginPage(WebDriver driver) {
+        PageFactory.initElements(driver, this);
+        this.driver = driver; }
+
+    @FindBy(id="user-name")
+    private WebElement loginField;
+
+    @FindBy(css = "input#login-button")
+    private WebElement loginBtn;
+
+    @FindBy(id = "password")
+    private WebElement passwdField;
+
+    public void inputLogin(String login) {
+        loginField.sendKeys(login); }
+
+    public void inputPasswd(String passwd) {
+        passwdField.sendKeys(passwd); }
+
+    public void clickLoginBtn() {
+        loginBtn.click(); }
+}
